[FIX] UAASDEV-8477, don't call compute_init_balance on empty account list
When exporting a report for a period range that does not include an opening
period, the initial balance will be requested from the static balance table.
For ledger accounts with report type 'none' this is done separately. Because
static balances do not accept an empty list of accounts, this breaks the
partner balance report because this report does not contain accounts with this
report type. Therefore, a check needs to be implemented if there are such
accounts amongst the accounts that are being reported on.